services/market-data/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
kafka-python==2.0.2
pandas==2.1.3
numpy==1.25.2
yfinance==0.2.28
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2